Added logic and tests

+++ HealthAssured.Checkout/ICheckout.cs
@@ -4,5 +4,5 @@ public interface ICheckout
 {
     void Scan(string item);
 
-    int GetTotalPrice();
+    decimal GetTotalPrice();
 }
